moved post to matter function to core module
