+++ src/planscape/planscape/check_celery.py
@@ -2,24 +2,13 @@
 import re
 import subprocess
 from collections import defaultdict
-from typing import Dict, Optional, Tuple
+from typing import Dict, Tuple
 
-import requests
+from core.mattermost import post_to_mattermost
 from decouple import Config, RepositoryEnv
 
 config = Config(RepositoryEnv("../../.env"))
 ENV = config("ENV")
-# === Configuration ===
-MATTERMOST_WEBHOOK_URL: Optional[str] = config(
-    "MATTERMOST_WEBHOOK_URL",
-    default=None,
-    cast=str,
-)  # type: ignore
-MATTERMOST_CHANNEL: str = config(
-    "MATTERMOST_CHANNEL",
-    "#planscape-alerts-dev",
-    cast=str,
-)  # type: ignore
 
 CELERY_CMD = [
     "/home/planscape/.local/bin/celery",
@@ -76,18 +65,6 @@ def format_mattermost_message(worker_counts: Dict[str, int], total: int) -> str:
         )
 
     return "\n".join(lines)
-
-
-def post_to_mattermost(message: str) -> None:
-    if not MATTERMOST_WEBHOOK_URL:
-        return
-    payload = {
-        "channel": MATTERMOST_CHANNEL,
-        "text": message,
-    }
-    response = requests.post(MATTERMOST_WEBHOOK_URL, json=payload)
-    if response.status_code != 200:
-        raise Exception(f"Failed to send message to Mattermost: {response.text}")
 
 
 def main():
